+++ totalRP3/modules/dashboard/WhatsNewTabView.lua
@@ -39,6 +39,7 @@ local strhtml = TRP3_API.utils.str.toHTML;
 --- List of "What's new" segments to concatenate and display together when
 ---  this view is activated.
 local SEGMENTS = {
+	loc.WHATS_NEW_24_10,
 	loc.WHATS_NEW_24_9,
 	loc.WHATS_NEW_24_8,
 	loc.WHATS_NEW_24_7,